--- project/techmya/src/main/java/com/ohgiraffers/techmya/shop/ShopProductController.java
@@ -1,20 +1,0 @@
-package com.ohgiraffers.techmya.shop;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/shop")
-public class ShopProductController {
-
-    @GetMapping("products")
-    public String pageShopProducts(){
-        return "shop/products";
-    }
-
-    @GetMapping("order")
-    public String pageOrderProduct(){
-        return "shop/orderTest";
-    }
-}
