+++ project/techmya/src/main/resources/templates/common/shop/fragments/header.html
@@ -182,7 +182,7 @@
                                             </a>
                                         </li>
                                         <li>
-                                            <a href="cart.html">
+                                            <a href="/shop/order">
                                                 <svg class="cart">
                                                     <use xlink:href="#cart"></use>
                                                 </svg>
