+++ project/techmya/src/main/resources/templates/shop/orderTest.html
@@ -6,20 +6,79 @@
 </head>
 <body>
 <div class="product-order-container">
-  <h1>주문 페이지</h1>
-  <form id="orderForm" action="" method="post">
-    <div class="input-group">
-      <label for="username">아이디:</label>
-      <input type="text" id="username" name="username" required>
-      <div id="usernameFeedback" class="feedback"></div>
-    </div>
-    <div class="input-group">
-      <label for="password">비밀번호:</label>
-      <input type="password" id="password" name="password" required>
-    </div>
-    <button type="submit">로그인</button>
+  <h1>주문 인서트 테스트 페이지</h1>
+  <form action="/shop/cart" method="post">
+    <label>상품번호: </label><input type="text" name="productNo"><br>
+    <label>옵션번호: </label><input type="text" name="productOptionNo"><br>
+    <label>주문수량: </label><input type="text" name="detailOrderAmount"><br>
 
+    <input type="submit" value="저장"/>
   </form>
+
+  <hr>
+  <h4>상품 번호 검색 시 옵션 선택이 뜨도록 해야함</h4>
+  <label>카테고리 코드: </label>
+  <select name="categoryCode">
+    <option value="1">식사</option>
+    <option value="2">음료</option>
+    <option value="3">디저트</option>
+    <option value="4">한식</option>
+    <option value="5">중식</option>
+    <option value="6">일식</option>
+    <option value="7">퓨전</option>
+  </select><br>
+
+
+  <hr>
+  <h1>장바구니</h1>
+  <table>
+    <tr>
+      <th>상품번호</th>
+      <th>옵션번호</th>
+      <th>주문수량</th>
+    </tr>
+    <tr th:each="product : ${ productlist }">
+      <td th:text="${ product.productNo }"></td>
+      <td th:text="${ product.productOptionNo }"></td>
+      <td th:text="${ product.detailOrderAmount }"></td>
+    </tr>
+  </table>
+  <button id="order-button">주문하기</button>
+
+  <script>
+    document.getElementById('order-button').addEventListener('click', function() {
+      // 장바구니 항목을 가져옴
+      const cartItems = document.querySelectorAll('#cart-items li');
+      let cartData = [];
+
+      cartItems.forEach(item => {
+        cartData.push({
+          id: item.getAttribute('data-id'),
+          name: item.textContent
+        });
+      });
+
+      // 서버로 POST 요청 보냄
+      fetch('/order', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify(cartData)
+      })
+              .then(response => response.json())
+              .then(data => {
+                console.log('Success:', data);
+                alert('주문이 성공적으로 접수되었습니다.');
+              })
+              .catch((error) => {
+                console.error('Error:', error);
+                alert('주문 처리 중 오류가 발생했습니다.');
+              });
+    });
+  </script>
+
+
 </div>
 </body>
 </html>
